Reorganize to optimize context usage
